Fix get_user_first_group for user with no groups
If the email is defined in user-groups.yaml but has no groups, groups is
assigned None and len(groups) throws an exception.

Intuitively, one would expect groups to default to [] rather than None
because [] is passed as the second argument to Dictionary.get, but this
default only applies if the key is not found in the dictionary. In this
case the key is defined but has a value of None.